moved utility tests
